Moved other diff directories to ...main/diff/.
