+++ src/main/java/hudson/plugins/report/jck/main/diff/formatters/ColorFormatter.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.plugins.report.jck.main.formatters;
+package hudson.plugins.report.jck.main.diff.formatters;
 
 import java.io.PrintStream;
