+++ src/main/java/hudson/plugins/report/jck/main/diff/web/ContentReader.java
@@ -21,28 +21,48 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.plugins.report.jck.main.formatters;
+package hudson.plugins.report.jck.main.diff.web;
 
-import java.io.PrintStream;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 
-public class HtmlFormatter2 extends HtmlFormatter {
+/**
+ * Class to read and save content of stdout/stderr of process.
+ */
+class ContentReader implements Runnable {
 
-    public HtmlFormatter2(PrintStream stream) {
-        super(stream);
+    StringBuilder sb = new StringBuilder();
+    private final InputStream is;
+
+    public String getContent() {
+        return sb.toString();
+    }
+
+    public ContentReader(InputStream is) throws IOException {
+        this.is = is;
     }
 
     @Override
-    public void small() {
-        print("<small>");
-        clossingBuffer.add("</small>");
+    public void run() {
+        try {
+            runImpl();
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
-    @Override
-    public void closeBuildsList() {
-        reset();
-        small();
-        print(" |x| ");
-        reset();
+    private void runImpl() throws IOException {
+        try (BufferedReader br = new BufferedReader(new InputStreamReader(is, "UTF-8"))) {
+            while (true) {
+                String line = br.readLine();
+                if (line == null) {
+                    break;
+                }
+                sb.append(line).append("\n");
+            }
+        }
     }
 
 }
